+++ server/odc-migrate/src/main/resources/init-config/init/regulation-rule-applying.yaml
@@ -450,6 +450,7 @@
     - 'OB_MYSQL'
     - 'MYSQL'
     - 'OB_ORACLE'
+    - 'ORACLE'
   propertiesJson: '{"${com.oceanbase.odc.builtin-resource.regulation.rule.sql-check.restrict-sql-affected-rows.allowed-max-sql-affected-count}":1000}'
 - enabled: 1
   level: 0
@@ -952,6 +953,7 @@
     - 'OB_MYSQL'
     - 'MYSQL'
     - 'OB_ORACLE'
+    - 'ORACLE'
   propertiesJson: '{"${com.oceanbase.odc.builtin-resource.regulation.rule.sql-check.restrict-sql-affected-rows.allowed-max-sql-affected-count}":10000}'
 - enabled: 1
   level: 0
@@ -1454,6 +1456,7 @@
     - 'OB_MYSQL'
     - 'MYSQL'
     - 'OB_ORACLE'
+    - 'ORACLE'
   propertiesJson: '{"${com.oceanbase.odc.builtin-resource.regulation.rule.sql-check.restrict-sql-affected-rows.allowed-max-sql-affected-count}":1000}'
 - enabled: 1
   level: 1
@@ -1946,6 +1949,7 @@
     - 'OB_MYSQL'
     - 'MYSQL'
     - 'OB_ORACLE'
+    - 'ORACLE'
   propertiesJson: '{"${com.oceanbase.odc.builtin-resource.regulation.rule.sql-check.restrict-sql-affected-rows.allowed-max-sql-affected-count}":1000}'
 - enabled: 1
   level: 1
