+++ server/odc-migrate/src/main/resources/init-config/init/regulation-rule-metadata.yaml
@@ -1418,6 +1418,8 @@
       value: MYSQL
     - label: SUPPORTED_DIALECT_TYPE
       value: OB_ORACLE
+    - label: SUPPORTED_DIALECT_TYPE
+      value: ORACLE
   propertyMetadatas:
     - name: ${com.oceanbase.odc.builtin-resource.regulation.rule.sql-check.restrict-sql-affected-rows.allowed-max-sql-affected-count}
       description: ${com.oceanbase.odc.builtin-resource.regulation.rule.sql-check.restrict-sql-affected-rows.allowed-max-sql-affected-count.description}
